AF Connect Once conscent flow: Button navigation
This is a version that divides the consent step
into two views navigated by two buttons
representing Step 1 - "Terms"
and Step 2 - "Consent checkboxes".

This is a suggestion how to
assure that the user
arrive at the desired point
in the UI for consent completion.

By implementing distinct steps
the user would not miss the
key actions to complete.

*